STM32 MBEDTLS_ALT use singleton
