+++ features/mbedtls/targets/TARGET_STM/md5_alt.cpp
@@ -15,95 +15,85 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  *
- *  This file implements STMicroelectronics SHA1 with HW services based on API
+ *  This file implements STMicroelectronics MD5 with HW services based on API
  *  from mbed TLS
  */
 /*
- *  The SHA-1 standard was published by NIST in 1993.
+ *  The MD5 algorithm was designed by Ron Rivest in 1991.
  *
- *  http://www.itl.nist.gov/fipspubs/fip180-1.htm
+ *  http://www.ietf.org/rfc/rfc1321.txt
  */
 
 /* Includes ------------------------------------------------------------------*/
-#include "mbedtls/sha1.h"
+#include "mbedtls/md5.h"
 
-#if defined(MBEDTLS_SHA1_C)
-#if defined(MBEDTLS_SHA1_ALT)
+#if defined(MBEDTLS_MD5_C)
+#if defined(MBEDTLS_MD5_ALT)
 #include <string.h>
 #include "mbedtls/platform.h"
 #include "mbedtls/platform_util.h"
 
+#include "platform/PlatformMutex.h"
+#include "platform/SingletonPtr.h"
+
+static SingletonPtr<PlatformMutex> md5_mutex;
 
 /* Private typedef -----------------------------------------------------------*/
 /* Private define ------------------------------------------------------------*/
 /* Private macro -------------------------------------------------------------*/
-#define SHA1_VALIDATE_RET(cond)                             \
-    MBEDTLS_INTERNAL_VALIDATE_RET( cond, MBEDTLS_ERR_SHA1_BAD_INPUT_DATA )
-#define SHA1_VALIDATE(cond)  MBEDTLS_INTERNAL_VALIDATE( cond )
+#define MD5_VALIDATE_RET(cond)                             \
+    MBEDTLS_INTERNAL_VALIDATE_RET( cond, MBEDTLS_ERR_MD5_BAD_INPUT_DATA )
+#define MD5_VALIDATE(cond)  MBEDTLS_INTERNAL_VALIDATE( cond )
 
 /* Private variables ---------------------------------------------------------*/
 /* Private function prototypes -----------------------------------------------*/
 /* Private functions ---------------------------------------------------------*/
 
-void mbedtls_sha1_init(mbedtls_sha1_context *ctx)
+void mbedtls_md5_init(mbedtls_md5_context *ctx)
 {
-    SHA1_VALIDATE(ctx != NULL);
+    MD5_VALIDATE(ctx != NULL);
 
-    __disable_irq();
-#if defined(MBEDTLS_THREADING_C)
-    /* mutex cannot be initialized twice */
-    if (!hash_mutex_started) {
-        mbedtls_mutex_init(&hash_mutex);
-        hash_mutex_started = 1;
-    }
-#endif /* MBEDTLS_THREADING_C */
+    md5_mutex->lock();
     hash_context_count++;
-    __enable_irq();
+    md5_mutex->unlock();
 
-    hash_zeroize(ctx, sizeof(mbedtls_sha1_context));
+    hash_zeroize(ctx, sizeof(mbedtls_md5_context));
 }
 
-void mbedtls_sha1_free(mbedtls_sha1_context *ctx)
+void mbedtls_md5_free(mbedtls_md5_context *ctx)
 {
     if (ctx == NULL) {
         return;
     }
 
-    __disable_irq();
+    md5_mutex->lock();
     if (hash_context_count > 0) {
         hash_context_count--;
-    }
 
-#if defined(MBEDTLS_THREADING_C)
-    if (hash_mutex_started) {
-        mbedtls_mutex_free(&hash_mutex);
-        hash_mutex_started = 0;
+        /* Shut down HASH on last context */
+        if (hash_context_count == 0) {
+            HAL_HASH_DeInit(&ctx->hhash);
+        }
     }
-#endif /* MBEDTLS_THREADING_C */
-    __enable_irq();
+    md5_mutex->unlock();
 
-    /* Shut down HASH on last context */
-    if (hash_context_count == 0) {
-        HAL_HASH_DeInit(&ctx->hhash);
-    }
-
-    hash_zeroize(ctx, sizeof(mbedtls_sha1_context));
+    hash_zeroize(ctx, sizeof(mbedtls_md5_context));
 }
 
-void mbedtls_sha1_clone(mbedtls_sha1_context *dst,
-                        const mbedtls_sha1_context *src)
+void mbedtls_md5_clone(mbedtls_md5_context *dst,
+                       const mbedtls_md5_context *src)
 {
-    SHA1_VALIDATE(dst != NULL);
-    SHA1_VALIDATE(src != NULL);
+    MD5_VALIDATE(dst != NULL);
+    MD5_VALIDATE(src != NULL);
 
     *dst = *src;
 }
 
-int mbedtls_sha1_starts_ret(mbedtls_sha1_context *ctx)
+int mbedtls_md5_starts_ret(mbedtls_md5_context *ctx)
 {
     int ret = 0;
 
-    SHA1_VALIDATE_RET(ctx != NULL);
+    MD5_VALIDATE_RET(ctx != NULL);
 
 #if defined(MBEDTLS_THREADING_C)
     if ((ret = mbedtls_mutex_lock(&hash_mutex)) != 0) {
@@ -136,13 +126,13 @@ exit :
     return (ret);
 }
 
-int mbedtls_internal_sha1_process(mbedtls_sha1_context *ctx,
-                                  const unsigned char data[ST_SHA1_BLOCK_SIZE])
+int mbedtls_internal_md5_process(mbedtls_md5_context *ctx,
+                                 const unsigned char data[ST_MD5_BLOCK_SIZE])
 {
     int ret = 0;
 
-    SHA1_VALIDATE_RET(ctx != NULL);
-    SHA1_VALIDATE_RET((const unsigned char *)data != NULL);
+    MD5_VALIDATE_RET(ctx != NULL);
+    MD5_VALIDATE_RET((const unsigned char *)data != NULL);
 
 #if defined(MBEDTLS_THREADING_C)
     if ((ret = mbedtls_mutex_lock(&hash_mutex)) != 0) {
@@ -153,9 +143,9 @@ int mbedtls_internal_sha1_process(mbedtls_sha1_context *ctx,
     /* restore hw context */
     HAL_HASH_ContextRestoring(&ctx->hhash, (uint8_t *)ctx->ctx_save_regs);
 
-    if (HAL_HASH_SHA1_Accmlt(&ctx->hhash,
-                             (uint8_t *) data,
-                             ST_SHA1_BLOCK_SIZE) != 0) {
+    if (HAL_HASH_MD5_Accmlt(&ctx->hhash,
+                            (uint8_t *) data,
+                            ST_MD5_BLOCK_SIZE) != 0) {
         ret = MBEDTLS_ERR_PLATFORM_HW_ACCEL_FAILED;
         goto exit;
     }
@@ -174,15 +164,15 @@ exit :
     return (ret);
 }
 
-int mbedtls_sha1_update_ret(mbedtls_sha1_context *ctx,
-                            const unsigned char *input,
-                            size_t ilen)
+int mbedtls_md5_update_ret(mbedtls_md5_context *ctx,
+                           const unsigned char *input,
+                           size_t ilen)
 {
     int ret = 0;
     size_t currentlen = ilen;
 
-    SHA1_VALIDATE_RET(ctx != NULL);
-    SHA1_VALIDATE_RET(ilen == 0 || input != NULL);
+    MD5_VALIDATE_RET(ctx != NULL);
+    MD5_VALIDATE_RET(ilen == 0 || input != NULL);
 
 #if defined(MBEDTLS_THREADING_C)
     if ((ret = mbedtls_mutex_lock(&hash_mutex)) != 0) {
@@ -193,39 +183,39 @@ int mbedtls_sha1_update_ret(mbedtls_sha1_context *ctx,
     /* restore hw context */
     HAL_HASH_ContextRestoring(&ctx->hhash, (uint8_t *)ctx->ctx_save_regs);
 
-    if (currentlen < (ST_SHA1_BLOCK_SIZE - ctx->sbuf_len)) {
+    if (currentlen < (ST_MD5_BLOCK_SIZE - ctx->sbuf_len)) {
         /* only store input data in context buffer */
         memcpy(ctx->sbuf + ctx->sbuf_len, input, currentlen);
         ctx->sbuf_len += currentlen;
     } else {
-        /* fill context buffer until ST_SHA1_BLOCK_SIZE bytes, and process it */
+        /* fill context buffer until ST_MD5_BLOCK_SIZE bytes, and process it */
         memcpy(ctx->sbuf + ctx->sbuf_len,
                input,
-               (ST_SHA1_BLOCK_SIZE - ctx->sbuf_len));
-        currentlen -= (ST_SHA1_BLOCK_SIZE - ctx->sbuf_len);
+               (ST_MD5_BLOCK_SIZE - ctx->sbuf_len));
+        currentlen -= (ST_MD5_BLOCK_SIZE - ctx->sbuf_len);
 
-        if (HAL_HASH_SHA1_Accmlt(&ctx->hhash,
-                                 (uint8_t *)(ctx->sbuf),
-                                 ST_SHA1_BLOCK_SIZE) != 0) {
+        if (HAL_HASH_MD5_Accmlt(&ctx->hhash,
+                                (uint8_t *)(ctx->sbuf),
+                                ST_MD5_BLOCK_SIZE) != 0) {
             ret = MBEDTLS_ERR_PLATFORM_HW_ACCEL_FAILED;
             goto exit;
         }
 
         /* Process following input data
-                      with size multiple of ST_SHA1_BLOCK_SIZE bytes */
-        size_t iter = currentlen / ST_SHA1_BLOCK_SIZE;
+                      with size multiple of ST_MD5_BLOCK_SIZE bytes */
+        size_t iter = currentlen / ST_MD5_BLOCK_SIZE;
         if (iter != 0) {
-            if (HAL_HASH_SHA1_Accmlt(&ctx->hhash,
-                                     (uint8_t *)(input + ST_SHA1_BLOCK_SIZE - ctx->sbuf_len),
-                                     (iter * ST_SHA1_BLOCK_SIZE)) != 0) {
+            if (HAL_HASH_MD5_Accmlt(&ctx->hhash,
+                                    (uint8_t *)(input + ST_MD5_BLOCK_SIZE - ctx->sbuf_len),
+                                    (iter * ST_MD5_BLOCK_SIZE)) != 0) {
                 ret = MBEDTLS_ERR_PLATFORM_HW_ACCEL_FAILED;
                 goto exit;
             }
         }
 
         /* Store only the remaining input data
-                        up to (ST_SHA1_BLOCK_SIZE - 1) bytes */
-        ctx->sbuf_len = currentlen % ST_SHA1_BLOCK_SIZE;
+                        up to (ST_MD5_BLOCK_SIZE - 1) bytes */
+        ctx->sbuf_len = currentlen % ST_MD5_BLOCK_SIZE;
         if (ctx->sbuf_len != 0) {
             memcpy(ctx->sbuf, input + ilen - ctx->sbuf_len, ctx->sbuf_len);
         }
@@ -245,13 +235,12 @@ exit :
     return (ret);
 }
 
-int mbedtls_sha1_finish_ret(mbedtls_sha1_context *ctx,
-                            unsigned char output[32])
+int mbedtls_md5_finish_ret(mbedtls_md5_context *ctx, unsigned char output[32])
 {
     int ret = 0;
 
-    SHA1_VALIDATE_RET(ctx != NULL);
-    SHA1_VALIDATE_RET((unsigned char *)output != NULL);
+    MD5_VALIDATE_RET(ctx != NULL);
+    MD5_VALIDATE_RET((unsigned char *)output != NULL);
 
 #if defined(MBEDTLS_THREADING_C)
     if ((ret = mbedtls_mutex_lock(&hash_mutex)) != 0) {
@@ -264,11 +253,11 @@ int mbedtls_sha1_finish_ret(mbedtls_sha1_context *ctx,
 
     /* Last accumulation for pending bytes in sbuf_len,
                              then trig processing and get digest */
-    if (HAL_HASH_SHA1_Accmlt_End(&ctx->hhash,
-                                 ctx->sbuf,
-                                 ctx->sbuf_len,
-                                 output,
-                                 ST_HASH_TIMEOUT) != 0) {
+    if (HAL_HASH_MD5_Accmlt_End(&ctx->hhash,
+                                ctx->sbuf,
+                                ctx->sbuf_len,
+                                output,
+                                ST_HASH_TIMEOUT) != 0) {
         ret = MBEDTLS_ERR_PLATFORM_HW_ACCEL_FAILED;
         goto exit;
     }
@@ -286,5 +275,5 @@ exit :
     return (ret);
 }
 
-#endif /* MBEDTLS_SHA1_ALT*/
-#endif /* MBEDTLS_SHA1_C */
+#endif /* MBEDTLS_MD5_ALT*/
+#endif /* MBEDTLS_MD5_C */
